# jobnow-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JobNow API",
        "description": "API documentation for JobNow - Moroccan Job Portal Platform. This API provides endpoints for job seekers and companies to manage job offers, applications, and user profiles.",
        "contact": {
            "name": "JobNow Support",
            "email": "support@jobnow.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.jobnow.ma",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/entreprise/analytics": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get entreprise analytics dashboard",
                "description": "Returns comprehensive analytics for the authenticated entreprise",
                "operationId": "e660f1286a9567b5acf8ea79f53555b2",
                "responses": {
                    "200": {
                        "description": "Analytics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidat/analytics": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get candidat analytics dashboard",
                "description": "Returns comprehensive analytics for the authenticated candidat",
                "operationId": "363baf6490e45e610a78ce9e84509adb",
                "responses": {
                    "200": {
                        "description": "Analytics retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}